+++ WidgetExtension/Up Next/UpNextLockScreenWidget.swift
@@ -9,7 +9,7 @@ struct UpNextLockScreenWidget: Widget {
             }
             .configurationDisplayName(L10n.upNext)
             .description(L10n.widgetsUpNextDescription)
-            .supportedFamilies([.accessoryCircular])
+            .supportedFamilies([.accessoryCircular, .accessoryRectangular])
         }
         else {
             return EmptyWidgetConfiguration()
@@ -20,6 +20,27 @@ struct UpNextLockScreenWidget: Widget {
 @available(iOSApplicationExtension 16.0, *)
 struct UpNextLockScreenWidgetEntryView: View {
     @State var entry: UpNextProvider.Entry
+    @Environment(\.widgetFamily) private var family
+
+    var body: some View {
+        switch family {
+        case .accessoryCircular:
+            UpNextCircularWidgetView(entry: entry)
+
+        case .accessoryRectangular:
+            UpNextRectangularWidgetView(entry: entry)
+
+        default:
+            EmptyView()
+        }
+    }
+}
+
+// MARK: - Circular View
+
+@available(iOSApplicationExtension 16.0, *)
+struct UpNextCircularWidgetView: View {
+    let entry: UpNextEntry
 
     var numberOfEpisodeInUpNext: Int {
         entry.upNextEpisodesCount ?? 0
@@ -32,7 +53,7 @@ struct UpNextLockScreenWidgetEntryView: View {
     var font: Font {
         numberOfEpisodeInUpNext > 99 ? .callout : .title
     }
-
+    
     var body: some View {
         ZStack {
             AccessoryWidgetBackground()
@@ -51,6 +72,65 @@ struct UpNextLockScreenWidgetEntryView: View {
             }
         }
         .widgetURL(URL(string: widgetURL))
+    }
+}
+
+// MARK: - Rectangle View
+
+@available(iOSApplicationExtension 16.0, *)
+struct UpNextRectangularWidgetView: View {
+    let entry: UpNextEntry
+
+    var nextEpisode: WidgetEpisode? {
+        // The first item returned is the currently playing episode
+        // so we check if the queue has at least 2 episodes in it since we pull the second one in the queue
+        guard let episodes = entry.episodes, episodes.count > 1 else {
+            return nil
+        }
+
+        return episodes[1]
+    }
+
+    var title: String {
+        nextEpisode?.episodeTitle ?? L10n.upNextEmptyTitle
+    }
+
+    var subtitle: String {
+        nextEpisode?.podcastName ?? L10n.widgetsNowPlayingTapDiscover
+    }
+
+    var widgetURL: String {
+        return nextEpisode != nil ? "pktc://upnext" : "pktc://discover"
+    }
+
+    var body: some View {
+        HStack {
+            VStack(alignment: .leading, spacing: 3) {
+                HStack(alignment: .center) {
+                    Image("up-next")
+                        .resizable()
+                        .frame(width: 12, height: 12)
+
+                    Text(L10n.upNext)
+                        .font(.headline)
+                        .fontWeight(.semibold)
+                        .foregroundColor(Color.primary)
+                        .layoutPriority(1)
+                }
+
+                Text(title)
+                    .font(.body)
+                    .fontWeight(.medium)
+                    .foregroundColor(Color.primary)
+                    .layoutPriority(1)
+                    .lineLimit(2)
+
+                Text(subtitle)
+                    .font(.body)
+                    .fontWeight(.medium)
+                    .foregroundColor(Color.secondary)
+            }
+        }.widgetURL(URL(string: widgetURL))
     }
 }